+++ config/runtime/prod.exs
@@ -87,6 +87,14 @@ config :explorer, Explorer.Repo.PolygonZkevm,
   pool_size: 1,
   ssl: ExplorerConfigHelper.ssl_enabled?()
 
+# Configures ZkSync database
+config :explorer, Explorer.Repo.ZkSync,
+  url: System.get_env("DATABASE_URL"),
+  # actually this repo is not started, and its pool size remains unused.
+  # separating repos for different CHAIN_TYPE is implemented only for the sake of keeping DB schema update relevant to the current chain type
+  pool_size: 1,
+  ssl: ExplorerConfigHelper.ssl_enabled?()
+
 # Configures Rootstock database
 config :explorer, Explorer.Repo.RSK,
   url: System.get_env("DATABASE_URL"),
@@ -116,8 +124,6 @@ config :explorer, Explorer.Repo.Filecoin,
 # Configures Stability database
 config :explorer, Explorer.Repo.Stability,
   url: System.get_env("DATABASE_URL"),
-  # actually this repo is not started, and its pool size remains unused.
-  # separating repos for different CHAIN_TYPE is implemented only for the sake of keeping DB schema update relevant to the current chain type
   pool_size: 1,
   ssl: ExplorerConfigHelper.ssl_enabled?()